refactor (api-objects): implementing ResourceResponseFormats from Knora API - move vocabulary response in sub folder - export of submodules in index.ts - provide vocabularies service appwide
